+++ xknx/devices/sensor.py
@@ -67,7 +67,7 @@ class Sensor(Device):
 
     def process_group_response(self, telegram: Telegram) -> None:
         """Process incoming GroupValueResponse telegrams."""
-        self.sensor_value.process(telegram)
+        self.sensor_value.process(telegram, always_callback=self.always_callback)
 
     def unit_of_measurement(self) -> str | None:
         """Return the unit of measurement."""